refactor. use bigdecimal.

+++ backend/src/main/java/robert/db/entities/Fee.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "FEE")
@@ -12,7 +13,7 @@ import javax.persistence.*;
 public class Fee extends BasicEntity {
 
     @Column(nullable = false)
-    private Double payedFee;
+    private BigDecimal payedFee;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "USER_ID")
